server/requirements.txt
asgiref==3.5.2
attrs==22.1.0
black==22.12.0
click==8.1.3
colorama==0.4.6
Django==4.1.4
django-cors-headers==3.13.0
django-dotenv==1.4.2
djangorestframework==3.14.0
djangorestframework-camel-case==1.3.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.25.1
inflection==0.5.1
isort==5.10.1
jsonschema==4.17.3
mypy-extensions==0.4.3
pathspec==0.10.3
platformdirs==2.6.0
psycopg2==2.9.5
PyJWT==2.6.0
pyrsistent==0.19.2
pytz==2022.6
PyYAML==6.0
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.7
uritemplate==4.1.1
